# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
host: localhost:5072
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getCrewList:
    # binds a127 app logic to a route
    x-swagger-router-controller: crews_ctrl
    post:
      tags: ['Crew Module']
      description: Get Crew List
      summary: Get Crew List
      # used as the method name of the controller
      operationId: getCrewList
      parameters:
        - name: body
          in: body
          description: get Crew By Foremen ID
          schema:
           $ref: '#/definitions/getCrewList'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                   
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/addCrewsImages:
    # binds a127 app logic to a route
    x-swagger-router-controller: crews_ctrl
    post:
      tags: ['Crew Module']
      description: Add Crew Images
      summary: Add Crew Images
      # used as the method name of the controller
      operationId: addCrewsImages
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
        - binary
      parameters:
        - in: formData
          name: id
          description: Crew id
          type: string
          required: true                
        - in: formData
          name: file
          description: Add Crew Images
          type: file
          required: true                      
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                                              
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /api/addUpdateCrewsData:
    # binds a127 app logic to a route
    x-swagger-router-controller: crews_ctrl
    post:
      tags: ['Crew Module']
      description: Add Crew Data
      summary: Add Crew Data
      # used as the method name of the controller
      operationId: addUpdateCrewsData
      parameters:
        - name: body
          in: body
          description: job parameters - firstname,lastname,email
          schema:
           $ref: '#/definitions/addCrewsData'
          required: true        
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getCrewById/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: crews_ctrl
    get:
      tags: ['Crew Module']
      description: Get Crew By Id 
      summary:  Get Crew By Id 
      # used as the method name of the controller
      operationId: getCrewById      
      parameters:
        - name: id
          in: path
          description: Crew Id
          required: true
          type: string        
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                         
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deleteCrewsById/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: crews_ctrl
    delete:
      tags: ['Crew Module']
      description: Delete Crew By Id
      summary: Delete Crew By Id
      # used as the method name of the controller
      operationId: deleteCrewsById
      parameters:
        - name: id
          in: path
          description: crew id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string              
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  addCrewsData:
    type: object
    properties:
      firstname:
        type: string 
      lastname:
         type: string
      email:
         type: string
      parent_id:
         type: string
      _id:
         type: string
...
  getCrewList:
    type: object
    properties:
      page:
        type: string         
      count:
        type: string      
      parent_id:
        type: string 
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string